kitty and font downloads are skipped if theyre already present

+++ scripts/_terminals.sh
@@ -33,19 +33,19 @@ function configure_tmux() {
     return
   }
 
-  copy_file $GIT_REPOS_DIR/tmux ~/.tmux/plugins/catppuccin/tmux
+  copy_file "$GIT_REPOS_DIR"/tmux ~/.tmux/plugins/catppuccin/tmux
   status=$?
   if [ $status -ne 0 ]; then
     return
   fi
 
-  copy_file $GIT_REPOS_DIR/tpm ~/.tmux/plugins/
+  copy_file "$GIT_REPOS_DIR"/tpm ~/.tmux/plugins/
   status=$?
   if [ $status -ne 0 ]; then
     return
   fi
 
-  copy_file $TMUX_CONF_SRC $TMUX_CONF_DST
+  copy_file "$TMUX_CONF_SRC" $TMUX_CONF_DST
   status=$?
   if [ $status -ne 0 ]; then
     return
@@ -59,37 +59,42 @@ function configure_tmux() {
 function configure_kitty() {
   start_step_message "Configuring Kitty"
 
-  start_step_message "Pulling Installer" "substep"
-  kitty_installer=$(mktemp /tmp/kitty_installer.XXXXXX.sh) || {
-    error_message "Failed to create temp file for kitty installer"
-    return
-  }
-  trap 'rm -f "$kitty_installer"' RETURN
-  if ! curl https://sw.kovidgoyal.net/kitty/installer.sh -o "$kitty_installer"; then
-    error_message "Failed to pull kitty installer"
-    return
-  fi
+  if ! command -v kitty >/dev/null 2>&1; then
 
-  start_step_message "Running Installer" "substep"
-  if ! /bin/sh "$kitty_installer"; then
-    error_message "Failed to run kitty installer"
-    return
+    start_step_message "Pulling Installer" "substep"
+    kitty_installer=$(mktemp /tmp/kitty_installer.XXXXXX.sh) || {
+      error_message "Failed to create temp file for kitty installer"
+      return
+    }
+    trap 'rm -f "$kitty_installer"' RETURN
+    if ! curl https://sw.kovidgoyal.net/kitty/installer.sh -o "$kitty_installer"; then
+      error_message "Failed to pull kitty installer"
+      return
+    fi
+
+    start_step_message "Running Installer" "substep"
+    if ! /bin/sh "$kitty_installer"; then
+      error_message "Failed to run kitty installer"
+      return
+    fi
+  else
+    info_message "Kitty exists... skipping installation"
   fi
 
   mkdir -p $KITTY_DIR/themes
-  copy_file $KITTY_CONF_SRC $KITTY_CONF_DST
+  copy_file "$KITTY_CONF_SRC" "$KITTY_CONF_DST"
   status=$?
   if [ $status -ne 0 ]; then
     return
   fi
 
-  cp $GIT_REPOS_DIR/kitty-themes/themes/* $KITTY_DIR/themes/
+  cp "$GIT_REPOS_DIR"/kitty-themes/themes/* $KITTY_DIR/themes/
   status=$?
   if [ $status -ne 0 ]; then
     return
   fi
 
-  copy_file $KITTY_DIR/themes/Broadcast.conf $KITTY_DIR/current-theme.conf
+  copy_file "$KITTY_DIR"/themes/Broadcast.conf $KITTY_DIR/current-theme.conf
   status=$?
   if [ $status -ne 0 ]; then
     return
@@ -103,7 +108,7 @@ function configure_kitty() {
 function configure_alacritty() {
   start_step_message "Configuring Alacritty"
   mkdir -p $ALACRITTY_DIR
-  copy_file $ALACRITTY_CONF_SRC $ALACRITTY_CONF_DST
+  copy_file "$ALACRITTY_CONF_SRC" "$ALACRITTY_CONF_DST"
   status=$?
   if [ $status -ne 0 ]; then
     return
